Detect missing "junction"="roundabout" tag
This script may be used for detecting roundabouts in OSM that are missing the "junction"="roundabout" tag. You may have to change some table names in the script for it to work. The database I used for testing the script was downloaded from Geofabrik and loaded to the Postgressql server using this script: https://github.com/baditaflorin/osm-postgis-scripts/blob/master/scope.sh